Removes BoardWrapper
Why:

* BoardWrapper was a Phaser tile that contained a Board tile within it.
Not only did this complicate the hierarchy, but there were also design
issues that might have caused problems in the future

Since each border had a margin between them (given by the BoardWrapper),
math wouldn't be direct (i.e. the start of the second board would be the
equal to the width of the first plus the border). For example, moving
pieces between boards would be extremely complicated

This change addresses the need by:

* Removing BoardWrapper, using just Board. Some of the functionality of
BoardWrapper that was still needed was ported to Board