Adds our usual eslint config, and refactors BoardCtrl
I copied my eslint config from Cadoo. I believe it's one of the most
up-to-date ones we have.

I also refactored the instantiation of BoardCtrl. It was being
instantiated inside the Board itself. Not only doesn't this make much
logical sense, it was also not even being used inside there. Only the
Game object would call for it, so I moved the instantiation there